requirements-global.txt
# base requirements
numpy==2.2.6
opencv-python==4.11.0.86
pillow==12.1.1
imagesize==1.4.1 #for concept statistics
tqdm==4.67.1
PyYAML==6.0.2
huggingface-hub==0.34.4
scipy==1.15.3
matplotlib==3.10.3
av==16.1.0
yt-dlp  #no pinned version, frequently updated for compatibility with sites
scenedetect==0.6.7.1
parse==1.20.2
setuptools==81.0.0

# pytorch
accelerate==1.12.0
safetensors==0.7.0
tensorboard==2.20.0
pytorch-lightning==2.6.1

# diffusion models
-e git+https://github.com/huggingface/diffusers.git@da6718f#egg=diffusers
gguf==0.17.1
transformers==4.57.6
sentencepiece==0.2.1 # transitive dependency of transformers for tokenizer loading
omegaconf==2.3.0 # needed to load stable diffusion from single ckpt files
invisible-watermark==0.2.0 # needed for the SDXL pipeline

# other models
pooch==1.8.2
open-clip-torch==2.32.0

# data loader
-e git+https://github.com/Calamdor/mgds.git@d278e09#egg=mgds

# optimizers
dadaptation==3.2 # dadaptation optimizers
lion-pytorch==0.2.3 # lion optimizer
prodigyopt==1.1.2 # prodigy optimizer
schedulefree==1.4.1 # schedule-free optimizers
pytorch_optimizer==3.6.0 # pytorch optimizers
prodigy-plus-schedule-free==2.0.1 # Prodigy plus optimizer
adv_optm==2.2.3 # advanced optimizers
-e git+https://github.com/KellerJordan/Muon.git@f90a42b#egg=muon-optimizer

# Profiling
scalene==1.5.51

# ui
customtkinter==5.2.2

# cloud
runpod==1.7.10
fabric==3.2.2

# debug
psutil==7.0.0
requests==2.33.0
deepdiff==8.6.2 # output easy to read diff for troublshooting
